Switched to using more streamlined and simplified panache mongo interface
