+++ BaseCode/Stats/build.gradle
@@ -4,7 +4,7 @@ plugins {
     id "io.freefair.lombok" version "4.1.5"
 }
 
-version = '1.3.0.128'
+version = '1.3.0.136'
 sourceCompatibility = 1.8
 targetCompatibility = 1.8
