+++ WebServer/Server/src/main/java/com/gjs/taskTimekeeper/webServer/server/mongoEntities/pojo/NotificationSettings.java
@@ -1,22 +1,18 @@
-package com.gjs.taskTimekeeper.webServer.server.pojo;
+package com.gjs.taskTimekeeper.webServer.server.mongoEntities.pojo;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.AllArgsConstructor;
-import lombok.Getter;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
+@Data
 @AllArgsConstructor
-
-public enum MembershipLevel {
-    MEMBER(0),
-    ADMIN_VIEW(1),
-    ADMIN_EDIT(2),
-    ADMIN_GROUP(3);
-
-    @Getter
-    private int levelValue;
+@NoArgsConstructor
+public class NotificationSettings {
+    private boolean notifyByEmail;
 
     public static void configureObjectMapper(ObjectMapper mapper){
-        //nothing to do
+        //nothing needs done
     }
 
     public static ObjectMapper getObjectMapper(){
