+++ WebServer/Server/src/main/java/com/gjs/taskTimekeeper/webServer/server/mongoEntities/pojo/UserLevel.java
@@ -1,0 +1,6 @@
+package com.gjs.taskTimekeeper.webServer.server.mongoEntities.pojo;
+
+public enum  UserLevel {
+    REGULAR,
+    ADMIN
+}
