--- WebServer/Server/src/main/java/com/gjs/taskTimekeeper/webServer/server/pojo/UserLevel.java
@@ -1,6 +1,0 @@
-package com.gjs.taskTimekeeper.webServer.server.pojo;
-
-public enum  UserLevel {
-    REGULAR,
-    ADMIN
-}
