+++ WebServer/WebLibrary/build.gradle
@@ -4,7 +4,7 @@ plugins {
     id "io.freefair.lombok" version "4.1.5"
 }
 
-version = '1.0.0.206'
+version = '1.0.0.214'
 sourceCompatibility = 1.8
 targetCompatibility = 1.8
